test-python-transcript-processing/ytbackend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
youtube-transcript-api>=1.0.0
yt-dlp
